xargs: Fix integer overflow in implicit conversion in do_exec.
* xargs/xargs.c (xargs_do_exec): Change type of r from int to
size_t, since that is the type that safe_read actually returns.

+++ xargs/xargs.c
@@ -1173,7 +1173,7 @@ xargs_do_exec (struct buildcmd_control *ctl, void *usercontext, int argc, char *
   pid_t child;
   int fd[2];
   int buf;
-  int r;
+  size_t r;
 
   (void) ctl;
   (void) argc;
